+++ lib/trailblazer/operation/pundit.rb
@@ -11,9 +11,7 @@ class Trailblazer::Operation
         Condition.new(*args, &block)
       end
 
-      # This can be subclassed for other policy strategies, e.g. non-pundit Authsome.
-      # NOTE: using a class here is faster than a simple proc: https://twitter.com/apotonick/status/791162989692891136
-      #
+      # Pundit::Condition is invoked at runtime when iterating the pipe.
       class Condition
         def initialize(policy_class, action)
           @policy_class, @action = policy_class, action
@@ -27,7 +25,7 @@ class Trailblazer::Operation
 
       private
         def build_policy(skills)
-          @policy_class.new(skills["user.current"], skills["model"])
+          @policy_class.new(skills["current_user"], skills["model"])
         end
 
         def result!(success, policy)