+++ test/operation/pundit_test.rb
@@ -27,7 +27,7 @@ class PolicyTest < Minitest::Spec
 
   # successful.
   it do
-    result = Create.({}, "user.current" => Module)
+    result = Create.({}, "current_user" => Module)
     result["process"].must_equal true
     #- result object, policy
     result["result.policy.default"].success?.must_equal true
@@ -37,14 +37,14 @@ class PolicyTest < Minitest::Spec
   end
   # breach.
   it do
-    result = Create.({}, "user.current" => nil)
+    result = Create.({}, "current_user" => nil)
     result["process"].must_equal nil
     #- result object, policy
     result["result.policy.default"].success?.must_equal false
     result["result.policy.default"]["message"].must_equal "Breach"
   end
-  # inject different policy.Condition  it { Create.({}, "user.current" => Object, "policy.default.eval" => Trailblazer::Operation::Policy::Pundit::Condition.new(Auth, :user_object?))["process"].must_equal true }
-  it { Create.({}, "user.current" => Module, "policy.default.eval" => Trailblazer::Operation::Policy::Pundit::Condition.new(Auth, :user_object?))["process"].must_equal nil }
+  # inject different policy.Condition  it { Create.({}, "current_user" => Object, "policy.default.eval" => Trailblazer::Operation::Policy::Pundit::Condition.new(Auth, :user_object?))["process"].must_equal true }
+  it { Create.({}, "current_user" => Module, "policy.default.eval" => Trailblazer::Operation::Policy::Pundit::Condition.new(Auth, :user_object?))["process"].must_equal nil }
 
 
   #---
@@ -57,7 +57,7 @@ class PolicyTest < Minitest::Spec
 
   # invalid because user AND model.
   it do
-    result = Show.({}, "user.current" => Module)
+    result = Show.({}, "current_user" => Module)
     result["process"].must_equal nil
     result["model"].inspect.must_equal %{#<struct PolicyTest::Song id=nil>}
     # result["policy"].inspect.must_equal %{#<struct PolicyTest::Song id=nil>}
@@ -66,7 +66,7 @@ class PolicyTest < Minitest::Spec
   # valid because new policy.
   it do
     # puts Show["pipetree"].inspect
-    result = Show.({}, "user.current" => Module, "policy.default.eval" => Trailblazer::Operation::Policy::Pundit::Condition.new(Auth, :user_and_model?))
+    result = Show.({}, "current_user" => Module, "policy.default.eval" => Trailblazer::Operation::Policy::Pundit::Condition.new(Auth, :user_and_model?))
     result["process"].must_equal true
     result["model"].inspect.must_equal %{#<struct PolicyTest::Song id=nil>}
     result["policy.default"].inspect.must_equal %{<Auth: user:Module, model:#<struct PolicyTest::Song id=nil>>}
@@ -84,7 +84,7 @@ class PolicyTest < Minitest::Spec
 
   # successful.
   it do
-    result = Edit.({ id: 1 }, "user.current" => Module)
+    result = Edit.({ id: 1 }, "current_user" => Module)
     result["process"].must_equal true
     result["model"].inspect.must_equal %{#<struct PolicyTest::Song id=1>}
     result["result.policy.default"].success?.must_equal true
@@ -95,7 +95,7 @@ class PolicyTest < Minitest::Spec
 
   # breach.
   it do
-    result = Edit.({ id: 4 }, "user.current" => nil)
+    result = Edit.({ id: 4 }, "current_user" => nil)
     result["model"].inspect.must_equal %{#<struct PolicyTest::Song id=4>}
     result["process"].must_equal nil
     result["result.policy.default"].success?.must_equal false