+++ CHANGES.md
@@ -35,6 +35,7 @@ TRB 1.2
 
 * The `model` method doesn't exist anymore, use `self["model"]` or write your own.
 * `find_by` diverts to left track.
+* `:create` is `:new` now.
 
 ## DI